Consolidated CalculatedRequiredLevyAmountExtensions and fixed tests
