+++ src/SFA.DAS.Payments.FundingSource.Application/Extensions/CalculatedRequiredLevyAmountExtensions.cs
@@ -1,4 +1,6 @@
 ﻿using System;
+using System.Collections.Generic;
+using System.Linq;
 using SFA.DAS.Payments.RequiredPayments.Messages.Events;
 
 namespace SFA.DAS.Payments.FundingSource.Application.Extensions
@@ -12,6 +14,26 @@ namespace SFA.DAS.Payments.FundingSource.Application.Extensions
             return message.TransferSenderAccountId.HasValue && 
                    message.TransferSenderAccountId != 0 && 
                    message.AccountId != message.TransferSenderAccountId;
+        }
+
+        public static List<CalculatedRequiredLevyAmount> WhereAndRemove(this List<CalculatedRequiredLevyAmount> sourceList, Func<CalculatedRequiredLevyAmount, bool> action)
+        {
+            var matchedItems = sourceList.Where(action).ToList();
+            matchedItems.ForEach(x => sourceList.Remove(x));
+            return matchedItems;
+        }
+
+        public static bool IsRefund(this CalculatedRequiredLevyAmount calculatedRequiredLevyAmount)
+        {
+            return calculatedRequiredLevyAmount.AmountDue < 0;
+        }
+
+        public static List<CalculatedRequiredLevyAmount> PerformDefaultPaymentsSort(this List<CalculatedRequiredLevyAmount> calculatedRequiredLevyAmounts)
+        {
+            return calculatedRequiredLevyAmounts
+                .OrderBy(x => x.AgreedOnDate)
+                .ThenBy(x => x.Learner.Uln)
+                .ToList();
         }
     }
 }
